Experimental changes to player
The player still uses inklecate behind the scenes-- however, the player view has been replaced by an iframe. To communicate between Inky and the iframe, a small API has been made in the form of bridge.js. The hope is eventually to support saving and loading projects and templates that each use their own bridge, allowing communication and debugging of a self contained HTML file and Javascript engine, which could then be zipped up without the bridge functions and played or published as-is.

These changes will allow easier development of InkJS projects for the web, and potentially even aid in the development of games for theoretical hybrid engines, such as Inksy (Bitsy + Ink, name not final), or even BiTwink (Bitsy + Twine + Ink, name final).